# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/health": {
      "get": {
        "operationId": "AppController_getHealth",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/api/clubs": {
      "get": {
        "operationId": "ClubsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClubResponse"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Clubs"
        ]
      }
    },
    "/api/clubs/{clubId}/members": {
      "get": {
        "operationId": "ClubsController_findMembers",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClubMemberResponse"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Clubs"
        ]
      }
    },
    "/api/clubs/{clubId}/books": {
      "get": {
        "operationId": "BooksController_list",
        "parameters": [
          {
            "name": "X-Member-Id",
            "in": "header",
            "description": "현재 멤버의 publicId — 쓰기 요청에 필요 (D-017)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "UPCOMING",
                "READING",
                "DONE"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedBooksResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Books"
        ]
      },
      "post": {
        "operationId": "BooksController_create",
        "parameters": [
          {
            "name": "X-Member-Id",
            "in": "header",
            "description": "현재 멤버의 publicId — 쓰기 요청에 필요 (D-017)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Books"
        ]
      }
    },
    "/api/books/{bookId}": {
      "get": {
        "operationId": "BooksController_detail",
        "parameters": [
          {
            "name": "X-Member-Id",
            "in": "header",
            "description": "현재 멤버의 publicId — 쓰기 요청에 필요 (D-017)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bookId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Books"
        ]
      },
      "patch": {
        "operationId": "BooksController_update",
        "parameters": [
          {
            "name": "X-Member-Id",
            "in": "header",
            "description": "현재 멤버의 publicId — 쓰기 요청에 필요 (D-017)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bookId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Books"
        ]
      },
      "delete": {
        "operationId": "BooksController_remove",
        "parameters": [
          {
            "name": "X-Member-Id",
            "in": "header",
            "description": "현재 멤버의 publicId — 쓰기 요청에 필요 (D-017)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bookId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Books"
        ]
      }
    },
    "/api/books/{bookId}/comments": {
      "get": {
        "operationId": "CommentsController_list",
        "parameters": [
          {
            "name": "X-Member-Id",
            "in": "header",
            "description": "현재 멤버의 publicId — 쓰기 요청에 필요 (D-017)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bookId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCommentsResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Comments"
        ]
      },
      "post": {
        "operationId": "CommentsController_create",
        "parameters": [
          {
            "name": "X-Member-Id",
            "in": "header",
            "description": "현재 멤버의 publicId — 쓰기 요청에 필요 (D-017)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bookId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Comments"
        ]
      }
    },
    "/api/comments/{commentId}": {
      "patch": {
        "operationId": "CommentsController_update",
        "parameters": [
          {
            "name": "X-Member-Id",
            "in": "header",
            "description": "현재 멤버의 publicId — 쓰기 요청에 필요 (D-017)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Comments"
        ]
      },
      "delete": {
        "operationId": "CommentsController_remove",
        "parameters": [
          {
            "name": "X-Member-Id",
            "in": "header",
            "description": "현재 멤버의 publicId — 쓰기 요청에 필요 (D-017)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Comments"
        ]
      }
    }
  },
  "info": {
    "title": "북클럽 로그 API",
    "description": "모임 책방·토론·문집 주문 API. 쓰기 요청은 X-Member-Id 헤더(멤버 publicId)로 현재 멤버를 식별한다.",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "ClubResponse": {
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "memberCount": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "publicId",
          "name",
          "description",
          "memberCount",
          "createdAt"
        ]
      },
      "ClubMemberResponse": {
        "type": "object",
        "properties": {
          "role": {
            "enum": [
              "LEADER",
              "MEMBER"
            ],
            "type": "string"
          },
          "publicId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "avatarEmoji": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "joinedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "role",
          "publicId",
          "name",
          "avatarEmoji",
          "color",
          "joinedAt"
        ]
      },
      "MemberSummaryResponse": {
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "avatarEmoji": {
            "type": "string"
          },
          "color": {
            "type": "string"
          }
        },
        "required": [
          "publicId",
          "name",
          "avatarEmoji",
          "color"
        ]
      },
      "BookResponse": {
        "type": "object",
        "properties": {
          "status": {
            "enum": [
              "UPCOMING",
              "READING",
              "DONE"
            ],
            "type": "string"
          },
          "publicId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "publisher": {
            "type": "string",
            "nullable": true
          },
          "coverColor": {
            "type": "string"
          },
          "coverEmoji": {
            "type": "string"
          },
          "meetingDate": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "periodFrom": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "periodTo": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberSummaryResponse"
            }
          },
          "commentCount": {
            "type": "number"
          }
        },
        "required": [
          "status",
          "publicId",
          "title",
          "author",
          "publisher",
          "coverColor",
          "coverEmoji",
          "meetingDate",
          "periodFrom",
          "periodTo",
          "createdAt",
          "updatedAt",
          "participants",
          "commentCount"
        ]
      },
      "PageMetaResponse": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "totalCount": {
            "type": "number"
          },
          "hasNext": {
            "type": "boolean"
          }
        },
        "required": [
          "page",
          "limit",
          "totalCount",
          "hasNext"
        ]
      },
      "PaginatedBooksResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookResponse"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PageMetaResponse"
          }
        },
        "required": [
          "items",
          "meta"
        ]
      },
      "CreateBookDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "author": {
            "type": "string",
            "maxLength": 100
          },
          "publisher": {
            "type": "string",
            "nullable": true,
            "maxLength": 100
          },
          "coverColor": {
            "type": "string",
            "pattern": "^#[0-9a-fA-F]{6}$"
          },
          "coverEmoji": {
            "type": "string",
            "maxLength": 16
          },
          "status": {
            "type": "string",
            "enum": [
              "UPCOMING",
              "READING",
              "DONE"
            ]
          },
          "meetingDate": {
            "type": "string",
            "nullable": true
          },
          "periodFrom": {
            "type": "string",
            "nullable": true
          },
          "periodTo": {
            "type": "string",
            "nullable": true
          },
          "participantIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "author",
          "coverColor",
          "coverEmoji"
        ]
      },
      "UpdateBookDto": {
        "type": "object",
        "properties": {}
      },
      "DeletedResponse": {
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          }
        },
        "required": [
          "publicId",
          "deleted"
        ]
      },
      "CommentResponse": {
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          },
          "member": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/MemberSummaryResponse"
              }
            ]
          },
          "page": {
            "type": "number",
            "nullable": true
          },
          "quote": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "isEdited": {
            "type": "boolean"
          }
        },
        "required": [
          "publicId",
          "deleted",
          "member",
          "page",
          "quote",
          "content",
          "createdAt",
          "updatedAt",
          "isEdited"
        ]
      },
      "CommentThreadResponse": {
        "type": "object",
        "properties": {
          "replies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentResponse"
            }
          },
          "publicId": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          },
          "member": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/MemberSummaryResponse"
              }
            ]
          },
          "page": {
            "type": "number",
            "nullable": true
          },
          "quote": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "isEdited": {
            "type": "boolean"
          }
        },
        "required": [
          "replies",
          "publicId",
          "deleted",
          "member",
          "page",
          "quote",
          "content",
          "createdAt",
          "updatedAt",
          "isEdited"
        ]
      },
      "PaginatedCommentsResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentThreadResponse"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PageMetaResponse"
          }
        },
        "required": [
          "items",
          "meta"
        ]
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 10000
          },
          "page": {
            "type": "number",
            "nullable": true,
            "minimum": 1
          },
          "quote": {
            "type": "string",
            "nullable": true,
            "maxLength": 1000
          },
          "parentId": {
            "type": "string"
          }
        },
        "required": [
          "content"
        ]
      },
      "UpdateCommentDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}